Create listComp.js
This is a short Javascript file I coded up to give users access to a function similar to that of a Python list comprehension. It will create Lists of all the various combinations of items of all the lists you give it. It will not modify the data inside of them in any way. If you are looking to do that consider using javascript .map function!